fix(composables): problem with reference to store in composables package
