+++ packages/composables/src/index.ts
@@ -10,22 +10,4 @@ export * from "./hooks/useUser";
 export * from "./hooks/useProductListing";
 export * from "./theme/cart/useCartSidebar";
 export * from "./theme/user/useUserLoginModal";
-
-/**
- * Workaround for current reactivity problems with SSR for Nuxt.
- * This section will be removed after Vuex is no longer in use.
- */
-let storeRef: any;
-
-/**
- * @alpha
- */
-export function setStore(ref: any) {
-  storeRef = ref;
-}
-/**
- * @alpha
- */
-export function getStore() {
-  return storeRef;
-}
+export * from "./store";
